muda o tipo para ponto flutuante

+++ src/main/java/dominio/banco/Banco.java
@@ -7,10 +7,10 @@ import dominio.bemImovel.BemImovel;
 import dominio.estruturaSocietaria.EstruturaSocietaria;
 
 public interface Banco {
-    static Long comprometimentoFinanceiro(EstruturaSocietaria estrutura){
+    static Double comprometimentoFinanceiro(EstruturaSocietaria estrutura){
         return estrutura.getSocios().stream()
             .flatMap(s -> s.getBens().stream())
             .map(b -> b.getValor())
-            .reduce(0L, Long::sum);
+            .reduce(0.0, Double::sum);
     }
 }
